Changed path handling and folder structure
Move the path specification to the config file.

Changed also how the temporary data is stored in the processed folder. It should be delivery specific and not version specific since the preparation steps are re-run if a new delivery comes in (and not necessarily if a new version should be generated).