Adjusted globals handling
Moved globals to the config file as well to avoid that this file (_targets) keeps changing when simply updating only e.g. the delivery.

Adjusted the global references in the respective functions as well.

Renamed the paths() reference to config_paths() to make sure that it is immediately clear that this part is referenced in the config file.