Adjusted strategy to handle empty columns
Some columns have 100% missings because they (presumingly) belong to residential RED. Before I dropped all columns that have 100% missings, but this could change between deliveries because might be adjusted/ reused. So, I defined a list of variables that should be removed in the process and stored it in the config file. This list can be expanded and adjusted if necessary.