Reworked how missing columns are handled
In the previous step, all missing columns are added to the data set. So the exemption for energieeffizienzklasse in Lieferung_2312 can be moved outside the if clause. Since it is now included in all data sets.

Which columns are suppose to be delete is now moved to the config file.

Completed the list of variables for the type transformation. Also build in a check if the variable is not existent, then it will be created with all missing values.