Handling missings in coordinates
Splitted the data preparation for observations that have a coordinate (the spatial joins will work) and those without a given coordinate (the spatial join will not work).

For observations with coordinate: Constructed municipality and district IDs by combining the given information (in gkz) and the information obtained from the spatial join. This should reduce the number of missings in these variables. The given information is prefered over the spatial join (assuming that people know what they are filling out).

For observations without coordinates: Adding the coordinate information as missing. Same holds for the grid information. Constructed the same municipality and district information out of the given information and merge the respective names using out Gebiete data.