Dependency and renaming of append_cleaned data
Made the housing_data_append_cleaned object depend on the previous step as the previous step saves a file to the local directory, it is not given that housing_data_append_cleaned is updated when the previous step changes (because R does not recognize the dependency between both steps).

Renamed the cleaning_append_data function to reflect the other naming styles.