Replace blid (stateID)
Replaced the blid with my own calculation based on the coordinates as the provided blid seems to be way off (Only very few observations are located in the same state when comparing my own blid and the provided blid). I also check some by hand and the provide blid was just wrong.