Removed rounding of coordinates
Removed the rounding of the coordinates because it is arbitrary and we do not have confirmation by ImmoScout that this is an issue.

The original explanation for doing this was that we would assume precision that is not given if we do not round.

See also project REDC Log issues:  #22012025-01 and #21012025-01