Test and change in variables for type setting
Implemented two tests that check that all presented variables are considered for the type setting step and that not variables are defined for type setting that will be deleted anyway.

Changed also the list of variables for the type setting. Some of them were missing and some of them weren't in the dataset. Therefore, also the checks.