Provider and security deposit were wrongly coded
The last option for provider was "unknown" which was coded as "9" but should be actually coded as missing (not specified) for consistency with other variables.

Security deposite type should be 6 (no deposit) when no deposit is required. However, this was masked by the other recoding in security deposit type. So, I moved it to the top to be the first decision.