Fixed deviations in blid
Based the blid on the district ID instead of the municipality ID. Shouldn't make a difference but should provide more stability as districts are less likely to change than municipalities.

For the dataset without coordinates, I still used the provided state ID (blid). So, I adjusted the strategy similar to the dataset with coordinates where the state ID is calculated based on the district ID.