Calculate end years only of microm is outdated
The function should only calculate the microm end years average if microm has fewer years than the housing data. This is, e.g., not the case in delivery 2312 as both datasets end in 2022. I added the condition here to make sure the calculation is only performed if housing data ends later.