Moved REDC version creation to append cleaning
The REDC version variable refers to the entire (complete) dataset after all the waves have been appended. Therefore, it is wrong to create the version variable in the cleaning of the raw data as this refers only to the newest wave. This step has been moved to cleaning the appended data with all the waves together.

Closes #40