Delete georeferencing_housing_data.R
Removed the function because its content got splitted into other functions mainly in adding_spatial_units.R. This is the advantage that the strategy is more modular and adding all the different spatial units does not happen in one function. If server fails, there is at least some progress saved.