Outsourced delivery label matching
Before you had to add a new line with new deliveries to label the redc_delivery variable accordingly. This triggered a change in the function which you would have needed to commit. I moved the recoding/ matching of deliveries and their labels to the globals and used dplyr::recode to avoid the adjustment in the function.