Create fixing_lmr_del_2306.R
Function to fix the LMR assignment for delivery 1 (2306). This is necessary because LMRs had been added as variable starting with version 2. However, since the process only cleans new incoming data and appends the previously cleaned data, the LMR had never been assigned to the cleaned version of delivery 1 (2306). Hence, the variable lmr2018 had been missing for all the observations in delivery 1 (2306). This function redoes the assignment only for delivery 1(2306) such that the variable exists for these observations downstream.